Split out error tests
